HR Analytics Case Study
A large company named XYZ , employs, at any given point of time, around 4000 employees. However, every year, around 15% of its employees leave the company and need to be replaced with the talent pool available in the job market. The management believes that this level of attrition (employees leaving, either on their own or because they got fired) is bad for the company, because of the following reasons
1. The former employees’ projects get delayed, which makes it difficult to meet timelines , resulting in a reputation loss among consumers and partners
2. A sizable department has to be maintained, for the purposes of recruiting new talent 
3. More often than not, the new employees have to be trained for the job and/or given time to acclimatize themselves to the company