Global Mart Sales Forecasting - Time Series
Business understanding:
“Global Mart” is an online store super giant having worldwide operations. It takes orders and delivers across the globe and deals with all the major product categories - consumer, corporate & home office.
Now as a sales/operations manager, you want to finalise the plan for the next 6 months.  So, you want to forecast the sales and the demand for the next 6 months, that would help you manage the revenue and inventory accordingly.
The store caters to 7 different market segments and in 3 major categories. You want to forecast at this granular level, so you subset your data into 21 (7*3) buckets before analysing these data.
But not all of these 21 market buckets are important from the store’s point of view. So you need to find out 2 most profitable (and consistent) segment from these 21 and forecast the sales and demand for these segments.